+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/rest/resources/ExperimentResource.java
@@ -1,5 +1,6 @@
 package edu.kit.ipd.crowdcontrol.objectservice.rest.resources;
 
+import edu.kit.ipd.crowdcontrol.objectservice.crowdworking.ExperimentOperator;
 import edu.kit.ipd.crowdcontrol.objectservice.crowdworking.PlatformManager;
 import edu.kit.ipd.crowdcontrol.objectservice.crowdworking.PreActionException;
 import edu.kit.ipd.crowdcontrol.objectservice.database.model.enums.ExperimentsPlatformModeMode;
